refactor: 포스트 컴포넌트 커스텀 훅스로 리팩토링
